+++ gateway/Dockerfile.build
@@ -10,7 +10,9 @@ RUN go get -d github.com/docker/docker/api/types \
 WORKDIR /go/src/github.com/alexellis/faas/gateway
 
 COPY metrics metrics
+COPY requests requests
 COPY server.go .
+COPY proxy.go .
 
 RUN find /go/src/github.com/alexellis/faas/gateway/
 RUN CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo -o app .
