+++ gateway/requests/requests.go
@@ -1,0 +1,23 @@
+package requests
+
+type AlexaSessionApplication struct {
+	ApplicationId string `json:"applicationId"`
+}
+
+type AlexaSession struct {
+	SessionId   string                  `json:"sessionId"`
+	Application AlexaSessionApplication `json:"application"`
+}
+
+type AlexaIntent struct {
+	Name string `json:"name"`
+}
+
+type AlexaRequest struct {
+	Intent AlexaIntent `json:"intent"`
+}
+
+type AlexaRequestBody struct {
+	Session AlexaSession `json:"session"`
+	Request AlexaRequest `json:"request"`
+}
