+++ prometheus/alert.rules
@@ -4,12 +4,12 @@ ALERT service_down
 ALERT APIHighInvocationRate
   IF rate ( gateway_function_invocation_total [10s] ) > 5
   FOR 30s
-  ANNOTATIONS {
-    summary = "High invocation total on {{ $labels.instance }}",
-    description =  "High invocation total on {{ $labels.instance }}",
-  }
   LABELS {
     service = "gateway",
     severity = "major",
-    value = "{{$value}}",
+    value = "{{$value}}"
   }
+  ANNOTATIONS {
+    summary = "High invocation total on {{ $labels.instance }}",
+    description =  "High invocation total on {{ $labels.instance }}"
+  }
